step_service/requirements.txt
fastapi==0.112.2
uvicorn==0.30.1
python-dotenv==1.0.0
httpx==0.28.1
boto3==1.35.0
python-multipart==0.0.6
celery[redis]==5.4.0
redis==7.2.0
trimesh==4.4.1
numpy==1.26.4
